add http put upload capability
An unauthenticated upload mechanism moves security concerns Elsewhere™,
but if carefully implemented keeps the team from having to be in the
secret-distribution business.

This is a nearly exact copy of the S3 presigned upload, with the
distinction that it includes the name of the archive in the upload
request (preventing having to specify a different URL per host).